changed price in all models from integerfield to decimalfield

+++ velocityapi/models/case.py
@@ -4,7 +4,7 @@ class Case(models.Model):
     title = models.CharField(max_length=100)
     link = models.CharField(max_length=100)
     img = models.CharField(max_length=100)
-    price = models.PositiveIntegerField()
+    price = models.DecimalField()
     brand = models.CharField(max_length=100)
     model = models.CharField(max_length=100)
     side_panel = models.CharField(max_length=100)
