+++ velocityapi/models/keyboard.py
@@ -4,7 +4,7 @@ class Keyboard(models.Model):
     title = models.CharField(max_length=100)
     link = models.CharField(max_length=100)
     img = models.CharField(max_length=100)
-    price = models.PositiveIntegerField()
+    price = models.DecimalField()
     brand = models.CharField(max_length=100)
     model = models.CharField(max_length=100)
     style = models.CharField(max_length=100)
